PC Assemble
On 10 November 2024, we listened to online talk from PETRONAS to get an overview of ICT jobs in the industry. The talk was very insightful as the speaker explained about PETRONAS Digital and description of the emerging jobs in ICT. Next time, I really hope that I can listen to the talk in-person as I have a lot of questions to ask to the speaker.